requirements.txt
flask
pyserial
numpy
wiringpi